refactor TextField
