+++ src/components/ui/primitives/text-field/index.ts
@@ -1,1 +1,1 @@
-export * from './textField'
+export * from './text-field'
