+++ src/components/ui/primitives/text-field/text-field.module.scss
@@ -1,20 +1,16 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-
-  .label,
-  ::placeholder {
-    color: var(--color-dark-300);
-  }
-
-  .label {
-    margin: 0;
-  }
+.label,
+::placeholder {
+  color: var(--color-dark-300); // был фикс типографики, удалить
 }
 
 .inputSvgWrapper {
   position: relative;
+  width: 100%;
+
+  &:focus-within {
+    color: inherit;
+    border: 1px solid var(--color-info-700);
+  }
 }
 
 .input {
@@ -28,20 +24,19 @@
 
   background-color: var(--color-dark-900);
   border: 1px solid var(--color-dark-300);
-  border-radius: var(--border-radius-xxs);
+  border-radius: var(--border-radius-xs);
+
+  &::placeholder {
+    font-style: italic;
+  }
 
   &:hover {
     background-color: var(--color-dark-700);
   }
 
   &:active {
-    color: var(--color-light-100);
+    // color: inherit;
     border: 1px solid var(--color-light-100);
-  }
-
-  &:focus,
-  &:focus-visible {
-    border: 2px solid var(--color-info-700);
   }
 
   &.error {
@@ -62,6 +57,16 @@
       outline: none;
     }
   }
+}
+
+.input:-webkit-autofill,
+.input:-webkit-autofill:hover,
+.input:-webkit-autofill:focus,
+.input:-webkit-autofill:active {
+  background-color: var(--color-dark-900);
+  transition: background-color 3000s ease-in-out 0s;
+
+  -webkit-text-fill-color: var(--color-light-100);
 }
 
 .icon {
@@ -92,8 +97,8 @@
 
 input[type='search'] {
   padding: 7px 30px;
-}
 
-input[type='search']::-webkit-search-cancel-button {
-  appearance: none;
+  &::-webkit-search-cancel-button {
+    appearance: none;
+  }
 }
